[lowpan-ot-driver] Force update of connectivity state after leave(+logs)
This change helps prevent a potential race condition regarding when the
internal `connectivity_state` variable is updated. Sometimes it was
being updated too slowly, which was causing subsequent calls to
`attach_all_nodes_to` to fail with `BadState`.

This change also adds some additional log lines to the error conditions
associated with `attach_all_nodes_to` to make sure that it is more clear
why it is failing when it fails.

Bug: b/296886787
Change-Id: I5fe55c2724e4f47a64b2d4fffb49aa295c16bf2e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/912935
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Commit-Queue: Robert Quattlebaum <rquattle@google.com>