[memory_sampler] De-duplicate stored stack traces
memory_sampler stores the stack traces of the allocations and
deallocations it records. Those tend to dominate the size, in memory,
of memory_sampler's records: we typically store an address, a size,
and a stack trace that can be more than a hundred frames deep.

We can save a substantial amount of memory in cases where the same few
functions perform many allocations, by reusing existing stack traces
and recording a reference instead.

Change-Id: Ibfccdf411f95407ca7b0090378cfdda23e162883
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/913673
Reviewed-by: Étienne J. Membrives <etiennej@google.com>
Reviewed-by: Fabio D'Urso <fdurso@google.com>
Commit-Queue: Guillaume Chelfi <chelfi@google.com>