[netstack3] Don't keep Ethernet devices in IdMap
Remove the load from ethernet devices' scalar identifier from the IdMap
backing. Use a HashMap instead, keyed on the allocation given by the rc
types.

Move the debug identifier to a named field in EthernetDeviceState to
prevent misuse.

The Weak version of the identifier copies the ID out on downgrade so
it's accessible when debugging, but it's not used in any equality or
other load-bearing comparisons or collection searching.

Extract the state inside Loopback and Ethernet states into type aliases
to avoid the long and hard to read repetition.

Change-Id: I7147ed440a505ba0bb4f5d997dd73705310a1031
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/913973
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>