[rtc][intel] Strip nvram support out of intel-rtc driver
The current intel-rtc driver has an improperly composed protocol of the
rtc and nvram protocols. The closed nature of the composed protocol is
making it difficult to do SDK-work on the rtc protocol at large. It
turns out that the nvram protocol isn't actually consumed by anything
other than the tests for the driver itself.

To remedy, I dropped the composed protocol and stripped out the nvram
interface of the intel-rtc driver. It is now just a provider of
ZX_PROTOCOL_RTC as per usual.

Fixes: 133214
Change-Id: I8f52b79f3980284211d43e1b6130891fb21f6489
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/913705
Commit-Queue: Sam Hansen <hansens@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
API-Review: David Pursell <dpursell@google.com>