[netstack3] Test interfaces-admin using TestSetup
The interfaces-admin implicit_address_removal_on_interface_removal test
directly constructs Ctx in order to test some address-removal behavior
when a device is removed. This relies on Ctx being Default-constructible
without using a NetstackSeed, which will imminently not be the case
after a subsequent CL.

This change makes that test use the integration-test TestSetup
fixtures instead. As part of this change, we switch to using a
netdevice rather than loopback interface, as the latter are never
removed during normal netstack operation. In order to support using
netdevice interfaces for tests like this where we need to spy on
the `InterfaceEvent`s produced, we introduce a TestSetup mechanism for
copying all `InterfaceEvent`s sent to the worker so that they are
observable from tests as well.

Multiply: netstack3-bin-instrumented
Change-Id: I8fdb8314e7b42bc65a73ec6efdb6545ffbe7b842
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/913152
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Fuchsia-Auto-Submit: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>