[superproject] Roll llvm-project/libc [libc] Implement stdio writing functions for the GPU port (#65809)
Summary:
This patch implements fwrite, putc, putchar, and fputc on the GPU. These
are very straightforward, the main difference for the GPU implementation
is that we are currently ignoring `errno`. This patch also introduces a
minimal smoke test for `putc` that is an exact copy of the `puts` test
except we print the string char by char. This also modifies the `fopen`
test to use `fwrite` to mirror its use of `fread` so that it is tested
as well.
GitOrigin-RevId: 8baf39bcfd11c6a12a6b77823e14f64ae121757d
Original-Revision: 80307938042d8edfc849663abe13c4deb983a8da
Roller-URL: https://ci.chromium.org/b/8770401383159187233
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I4daec6d519a20082b34c9139c86a043a1250ed56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/914356