[fidl][fuchsia.component.sandbox] Add to partner SDK
The reason for this is to make it possible to use sandbox types in other
fuchsia.component FIDL APIs.

Note that fuchsia.component.sandbox is versioned at HEAD, so the
namespace is still unavailable in the stable SDK -- that's important
since the APIs are still under development.

Bug: b/297412364, b/298100106
Change-Id: I1767cb7106a27b89002481b7e844079853e850bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/909553
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
API-Review: Gary Bressler <geb@google.com>