[netstack3] Prepare for Primary reference waiting
Simplify Primary a bit by removing the extra refs drop allowance.
Carrying that around was only necessary for `unwrap` and we can solve
that a bit more locally in `unwrap` without spreading the complexity
around.

Use ManuallyDrop inside Primary and Inner to allow unwrapping instead of
ptr_read. These are equivalent but ManuallyDrop is more purpose-built
for exactly this.

Add a test to check the double-panic path is protected and added a note
about unwind safety on Primary. The right thing to do here would
probably be to 'poison' the references and prevent further use, but we
don't need it.

In a follow-up, we'll allow unwrapping `data: T` from inner through a
notifier to close the loop on waiting for all references to drop.

Bug: 122388
Change-Id: I139a8959acae14ecb2715f2508496d101172fa54
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/914354
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>