[tefmocheck] Fix log splitting to match test names.
go/bbid/8770228455612327809 is an example where a line matched the
testFinishedRegex but wasn't the line we were looking for so it later
failed to process because the test name we parsed from that line didn't
match the test names in the summary.json. We should check that the test
name we parse from the line matches the expected test name; otherwise,
skip that line.

Change-Id: I99f9e5bca1afc1baceaf4c52e8064480a1e3ade9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/914956
Reviewed-by: Nathan Mulcahey <nmulcahey@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>