[ffx][pb] Change get-artifacts to also support emu
We are seeing infra bots failure when we try to use product bundle. The reason is, Product Bundles are too big for `test bots`, and we don't really need all the artifacts exist inside Product bundle directory to fulfill flashing/emulating of the test devices.

The solution we are trying is, in build bots, we invoke `ffx product get-artifacts` and upload only the necessary artifacts to CAS, so when test bots fetch all the artifacts from CAS, there will only be artifacts needed for flashing or emulating. This needs `ffx product get-artifacts`  to also support collecting emu artifacts.

Change-Id: I212d5680e850d8c80688cf372aaf67daf8d9ae29
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/913597
Commit-Queue: Jiaming Li <lijiaming@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>