[netstack3] Delete is_device_installed
The only place this was used was for eligibility for device selection
from routes. Disable IP on that device has the exact same effect, so
doing that on removal.

In the world where core stops trying to remove device routes by itself
and delegates this cleanup to bindings, this is likely irrelevant but
serves as a statement that as long as bindings itself doesn't attempt to
re-enable IPv[4|6] on that interface it shouldn't accept new addresses
or packets.

Change-Id: Ia2d11cb6a0546bb9fbb36925e7aaa422a471cbb0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/913901
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>