[ffx][emu] Improved error message when emulator not supported
This updates the error message to be more clear when the product bundle
does not contain virtual device specifications. These virtual devices
are used to configure the emulator. The are not present in product
bundles that do not support running in the emulator, such as product
bundles that use a physical board type that does contain emulation
support.

Bug: b/294271932
Change-Id: Ia85b28b4297499b775b32cf8ee4bbc1d83c82bb6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/915992
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>