[overnet] Improve ascendd socket error reporting
This CL improves the error reporting logic when failing to bind the
ascendd socket. Before, it would assume that bind() only fails when the
socket is already in use, and tried to connect to that existing socket.
Now, it only does that on EADDRINUSE, and reports the original bind()
error otherwise.

Change-Id: I7f662a36a4e52e5c7429680e3250d825489af6ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/916268
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Mitchell Kember <mkember@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>