[pw_rpc] Client/server component integration prototype
Add a prototype to //examples/components that demonstrates integration
between the Component Framework and an offloaded program that uses
pigweed.

The offloaded program is represented in Fuchsia as a component. This
component uses a 'pigweed runner' that bridges between Fuchsia and the
offloaded program (in this example, assumed to run on a Linux host).

The pigweed side change landed in
https://pigweed-review.git.corp.google.com/c/pigweed/pigweed/+/167158.

Bug: b/299358404

Change-Id: Idadc9bf2bbbe078b7b22120bdba1bd9e1553e7e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/878075
Reviewed-by: Joshua Conner <joshconner@google.com>
API-Review: Adam Barth <abarth@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Armando Montanez <amontanez@google.com>