[fuchsia-component] Support opening from a directory endpoint
Add support for opening protocols, files, directories, relative
to a directory endpoint, using fdio. This is useful when one only has
access to a &ClientEnd<DirectoryProxy>, hence cannot convert it to a
directory proxy.

fuchsia-fs is used to open these relative to a DirectoryProxy.
fxrev.dev/915818 was an attempt to add these to fuchsia-fs, but the
owners would prefer to avoid fdio in that library.

Change-Id: I162a9bc654c927ba5be8285199de73504118427e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917139
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>