[driver][sdk] Migrate to DriverRegistration
This change migrates the driver component libraries to use the FIDL
based DriverRegistration mechanism to export the driver. This involves
moving away from the legacy C lifecycle ABI.

There are also some changes needed in the compat driver as it was
using the CustomFactory internal feature which is not available
anymore. Instead the a server has been created just for the compat
driver so it can do the custom initialization logic it needed.

In order to not break existing drivers and force changes for them to
recompile, all of the customer facing APIs have been unchanged. Instead
the internal machinery has been changed to work with the new mechanism.
This includes the DriverBase methods users should overwrite and have
access to, the DriverBase constructor they should use, the testing
library APIs, the export macro usage, the async completers for start and
preparestop, etc...

Change-Id: I523a2ad2545b288e561f6f4c2a1ae06c81cc3837
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/891795
API-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>