[driver][sdk] Fix breaking Stop in test lib
In the migration cl to the driver registration, I made a breaking change
that changed the definition for the Stop call in the test library's
DriverLifecycle class. This reverts that breaking change. The
PrepareStop hook remains the entrypoint to the FIDL Stop to keep
consistent with the naming in DriverBase.

Bug: 133230
Change-Id: I112b13b74ba6fcefb46a601d0f50379c8c3c6bb7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918335
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>