[driver_runtime] Fix timer cancel bug.
Switch from async::Task to async_task_t.

async::Task is expected to run in a single-threaded environment,
so will assert if cancelation fails.

It's not a concern if cancelation fails, that means the timer just
woke up on another thread, and the handler is about to run.

Bug: 132992
Change-Id: Icd980a629690b353a259caec6e5b19b47132ab35
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/913893
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>