[starnix] Fix race on the binder driver
The previous implementation was releasing the same object multiple time
when an object had both strong and weak reference dropping to 0 in the
same cycle. That creates a race when the same object is added back
between the 2 removes and the new object, that is still references, is
then removed by the second removal.

This CL fixes this race by not removing the object immediatey. Instead
if keep track of all object that could be removed, and remove them at
the end a single time.

Fixed: b/298935909 b/291922706
Change-Id: I31c284689dbadf6ff77866ab1c6472e62412af18
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917973
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Barth <abarth@google.com>